+++ packages/fabric/editor-core/src/plugins/block-type/index.ts
@@ -21,7 +21,6 @@ import keymapHandler from './keymap';
 import inputRulePlugin from './input-rule';
 
 import { areBlockTypesDisabled } from '../../utils';
-import { EditorAppearance } from '../../editor/types/editor-props';
 
 export type StateChangeHandler = (state: BlockTypeState) => any;
 export type BlockTypeStateSubscriber = (state: BlockTypeState) => any;
@@ -169,32 +168,31 @@ export class BlockTypeState {
 
 export const stateKey = new PluginKey('blockTypePlugin');
 
-export const createPlugin = (appearance?: EditorAppearance | undefined) =>
-  new Plugin({
-    state: {
-      init(config, state: EditorState) {
-        return new BlockTypeState(state);
-      },
-      apply(tr, pluginState: BlockTypeState, oldState, newState) {
-        pluginState.update(newState);
-        return pluginState;
-      },
+export const plugin = new Plugin({
+  state: {
+    init(config, state: EditorState) {
+      return new BlockTypeState(state);
     },
-    key: stateKey,
-    view: (view: EditorView) => {
-      const pluginState = stateKey.getState(view.state);
-      pluginState.keymapHandler = keymapHandler(view, pluginState, appearance);
-      return {};
+    apply(tr, pluginState: BlockTypeState, oldState, newState) {
+      pluginState.update(newState);
+      return pluginState;
     },
-    props: {
-      handleKeyDown(view, event) {
-        return stateKey.getState(view.state).keymapHandler(view, event);
-      },
+  },
+  key: stateKey,
+  view: (view: EditorView) => {
+    const pluginState = stateKey.getState(view.state);
+    pluginState.keymapHandler = keymapHandler(view, pluginState);
+    return {};
+  },
+  props: {
+    handleKeyDown(view, event) {
+      return stateKey.getState(view.state).keymapHandler(view, event);
     },
-  });
+  },
+});
 
-const plugins = (schema: Schema, appearance?: EditorAppearance) => {
-  return [createPlugin(appearance), inputRulePlugin(schema)].filter(
+const plugins = (schema: Schema) => {
+  return [plugin, inputRulePlugin(schema)].filter(
     plugin => !!plugin,
   ) as Plugin[];
 };