+++ packages/fabric/editor-core/src/ui/Extension/index.tsx
@@ -2,10 +2,7 @@ import * as React from 'react';
 import { Component } from 'react';
 import { EditorView } from 'prosemirror-view';
 import { Node as PMNode } from 'prosemirror-model';
-import {
-  default as ProviderFactory,
-  WithProviders,
-} from '../../providerFactory';
+import { ProviderFactory, WithProviders } from '@atlaskit/editor-common';
 import ExtensionComponent from './ExtensionComponent';
 import {
   setExtensionElement,